+++ lib/mongoid/document.rb
@@ -1,5 +1,4 @@
 # encoding: utf-8
-require "mongoid/positional"
 require "mongoid/evolvable"
 require "mongoid/extensions"
 require "mongoid/errors"